flask_api/requirements.txt
Flask
pandas
joblib
google-cloud-storage
scikit-learn
gunicorn
